Create Week-3-3.3
n: In C++, the this pointer is an implicit pointer available inside all non-static member functions of a 
class. It points to the current object that invoked the member function. The this pointer is especially useful 
when you need to explicitly refer to the calling object’s members, such as resolving naming conflicts between 
function parameters and class data members. Using the this pointer improves code clarity and is essential in 
cases like method chaining or returning the current object from a member function.