Create Week-5-5.1
Constructor: A special function that is automatically called when an object is created. It's typically used to 
initialize member variables.  
Destructor: A special function that is automatically called when an object goes out of scope or is deleted. It's 
used to perform cleanup, such as releasing memory or closing files.