Create Week-5-5.3
: This program demonstrates the use of a copy constructor in the class Book.  
• The class Book has two constructors:  
o A parameterized constructor that initializes the book's title and page count.  
o A copy constructor that initializes a new Book object by copying the data from an existing 
Book object.  
• In the main() function:  
o book1 is created using the parameterized constructor.  
o book2 is created as a copy of book1, which calls the copy constructor.  
• Both objects' details are displayed using the display() method.